split interface/implementation for BlurHashDecoder
and limit the growth of the cosine caches
