+++ core/utils/src/commonMain/kotlin/com/livefast/eattrash/raccoonforfriendica/core/utils/di/Utils.kt
@@ -2,6 +2,7 @@ package com.livefast.eattrash.raccoonforfriendica.core.utils.di
 
 import com.livefast.eattrash.raccoonforfriendica.core.utils.debug.AppInfoRepository
 import com.livefast.eattrash.raccoonforfriendica.core.utils.gallery.GalleryHelper
+import com.livefast.eattrash.raccoonforfriendica.core.utils.imageload.BlurHashDecoder
 import com.livefast.eattrash.raccoonforfriendica.core.utils.imageload.ImageLoaderProvider
 import com.livefast.eattrash.raccoonforfriendica.core.utils.share.ShareHelper
 import org.koin.core.module.Module
@@ -13,6 +14,8 @@ expect fun getGalleryHelper(): GalleryHelper
 expect fun getShareHelper(): ShareHelper
 
 expect fun getAppInfoRepository(): AppInfoRepository
+
+expect fun getBlurHashDecoder(): BlurHashDecoder
 
 expect val coreUtilsFileSystemModule: Module
