+++ core/utils/src/commonMain/kotlin/com/livefast/eattrash/raccoonforfriendica/core/utils/di/UtilsModule.kt
@@ -1,5 +1,7 @@
 package com.livefast.eattrash.raccoonforfriendica.core.utils.di
 
+import com.livefast.eattrash.raccoonforfriendica.core.utils.imageload.BlurHashDecoder
+import com.livefast.eattrash.raccoonforfriendica.core.utils.imageload.DefaultBlurHashDecoder
 import com.livefast.eattrash.raccoonforfriendica.core.utils.imageload.DefaultImageLoaderProvider
 import com.livefast.eattrash.raccoonforfriendica.core.utils.imageload.DefaultImagePreloadManager
 import com.livefast.eattrash.raccoonforfriendica.core.utils.imageload.ImageLoaderProvider
@@ -20,5 +22,8 @@ val coreUtilsModule =
                 context = get(),
                 imageLoaderProvider = get(),
             )
+        }
+        single<BlurHashDecoder> {
+            DefaultBlurHashDecoder()
         }
     }
