+++ core/utils/src/iosMain/kotlin/com/livefast/eattrash/raccoonforfriendica/core/utils/di/Utils.kt
@@ -6,6 +6,7 @@ import com.livefast.eattrash.raccoonforfriendica.core.utils.fs.DefaultFileSystem
 import com.livefast.eattrash.raccoonforfriendica.core.utils.fs.FileSystemManager
 import com.livefast.eattrash.raccoonforfriendica.core.utils.gallery.DefaultGalleryHelper
 import com.livefast.eattrash.raccoonforfriendica.core.utils.gallery.GalleryHelper
+import com.livefast.eattrash.raccoonforfriendica.core.utils.imageload.BlurHashDecoder
 import com.livefast.eattrash.raccoonforfriendica.core.utils.imageload.ImageLoaderProvider
 import com.livefast.eattrash.raccoonforfriendica.core.utils.share.DefaultShareHelper
 import com.livefast.eattrash.raccoonforfriendica.core.utils.share.ShareHelper
@@ -24,6 +25,8 @@ actual fun getGalleryHelper(): GalleryHelper = CoreUtilsDiHelper.galleryHelper
 actual fun getShareHelper(): ShareHelper = CoreUtilsDiHelper.shareHelper
 
 actual fun getAppInfoRepository(): AppInfoRepository = CoreUtilsDiHelper.appInfoRepository
+
+actual fun getBlurHashDecoder(): BlurHashDecoder = CoreUtilsDiHelper.blurHashDecoder
 
 actual val coreUtilsFileSystemModule =
     module {
@@ -72,4 +75,5 @@ internal object CoreUtilsDiHelper : KoinComponent {
     val galleryHelper: GalleryHelper by inject()
     val shareHelper: ShareHelper by inject()
     val appInfoRepository: AppInfoRepository by inject()
+    val blurHashDecoder: BlurHashDecoder by inject()
 }
